Moved findFocusPath to glimmer-utils
